+++ src/api/apps/cruxql/dto/purchase-db.dto.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from "class-validator"
-
-export class PurchaseDbDto {
-  @IsNotEmpty()
-  apiKey: string
-
-  @IsNotEmpty()
-  dbId: string
-}