+++ src/api/subscription/subscription.service.ts
@@ -14,7 +14,6 @@ import { FrostlakeRepository } from "../apps/frostlake/frostlake.repository"
 import { DwalletRepository } from "../apps/dwallet/dwallet.repository"
 import { SwapstreamRepository } from "../apps/swapstream/swapstream.repository"
 import { SnowlakeRepository } from "../apps/snowlake/snowlake.repository"
-import { CruxqlRepository } from "../apps/cruxql/cruxql.repository"
 import { EasenftRepository } from "../apps/easenft/easenft.repository"
 import { VuelockRepository } from "../apps/vuelock/vuelock.repository"
 
@@ -30,7 +29,6 @@ export class SubscriptionService {
     private readonly frostlakeRepository: FrostlakeRepository,
     private readonly swapstreamRepository: SwapstreamRepository,
     private readonly snowlakeRepository: SnowlakeRepository,
-    private readonly cruxqlRepository: CruxqlRepository,
     private readonly easenftRepository: EasenftRepository,
     private readonly vuelockRepository: VuelockRepository) {
     this.infuraEndpoint = otherConstants.infuraEndpoint + "/" + envConfig.infuraApiKey
@@ -103,7 +101,6 @@ export class SubscriptionService {
       if (subscription) {
         const { apiKey } = subscription
         const airlakeUsedCredits = await this.airlakeRepository.findCountByApiKey(apiKey) * apiPricing.airlake
-        const cruxqlUsedCredits = await this.cruxqlRepository.findCountByApiKey(apiKey) * apiPricing.cruxql
         const dwalletUsedCredits = await this.dwalletRepository.findCountByApiKey(apiKey) * apiPricing.dwallet
         const easeNftUsedCredits = await this.easenftRepository.findCountByApiKey(apiKey) * apiPricing.easenft
         const frostlakeUsedCredits = await this.frostlakeRepository.findCountByApiKey(apiKey) * apiPricing.frostlake
@@ -112,7 +109,6 @@ export class SubscriptionService {
         const vuelockUsedCredits = await this.vuelockRepository.findCountByApiKey(apiKey) * apiPricing.vuelock
 
         const usedCredits = airlakeUsedCredits +
-          cruxqlUsedCredits +
           dwalletUsedCredits +
           easeNftUsedCredits +
           frostlakeUsedCredits +