+++ src/authorization/apikeyauthorizer/apikeyauthorizer.decorator.ts
@@ -1,6 +1,5 @@
 import { createParamDecorator, ExecutionContext, ForbiddenException } from "@nestjs/common"
 import { AirlakeHistoryModel } from "src/api/apps/airlake/entities/airlake-history.entity"
-import { CruxqlDbOwnershipModel } from "src/api/apps/cruxql/entities/cruxql-dbownership.entity"
 import { DwalletTransactionModel } from "src/api/apps/dwallet/entities/dwallet.entity"
 import { EasenftTransactionModel } from "src/api/apps/easenft/entities/easenft.entity"
 import { FrostlakeAnalyticsModel } from "src/api/apps/frostlake/entities/frostlake-analytics.entity"
@@ -36,7 +35,6 @@ export const ApiKeyAuthorizer = createParamDecorator(
 
           else {
             const airlakeUsedCredits = await AirlakeHistoryModel.find({ apiKey }).countDocuments() * apiPricing.airlake
-            const cruxqlUsedCredits = await CruxqlDbOwnershipModel.find({ apiKey }).countDocuments() * apiPricing.cruxql
             const dwalletUsedCredits = await DwalletTransactionModel.find({ apiKey }).countDocuments() * apiPricing.dwallet
             const easeNftUsedCredits = await EasenftTransactionModel.find({ apiKey }).countDocuments() * apiPricing.easenft
             const frostlakeUsedCredits = await FrostlakeAnalyticsModel.find({ apiKey }).countDocuments() * apiPricing.frostlake
@@ -44,7 +42,6 @@ export const ApiKeyAuthorizer = createParamDecorator(
             const swapstreamUsedCredits = await SwapstreamTransactionModel.find({ apiKey }).countDocuments() * apiPricing.swapstream
             const vuelockUsedCredits = await VuelockSecretModel.find({ apiKey }).countDocuments() * apiPricing.vuelock
             const usedCredits = airlakeUsedCredits +
-              cruxqlUsedCredits +
               dwalletUsedCredits +
               easeNftUsedCredits +
               frostlakeUsedCredits +