+++ src/config/envConfig.ts
@@ -6,7 +6,6 @@ export const envConfig = {
   apiPort: process.env.API_PORT,
   lenstackPlatformDbUri: process.env.LENSTACK_PLATFORM_DB_URI,
   airlakeMongoDbUri: process.env.AIRLAKE_DB_URI,
-  cruxqlMongoDbUri: process.env.CRUXQL_DB_URI,
   dwalletMongoDbUri: process.env.DWALLET_DB_URI,
   easenftMongoDbUri: process.env.EASENFT_DB_URI,
   frostlakeMongoDbUri: process.env.FROSTLAKE_DB_URI,