+++ src/config/subscriptionConfig.ts
@@ -1,21 +1,20 @@
 export const subscriptionConfig = {
   trialSubscriptionConfig: {
     price: 0,
-    grantedCredits: 15000
+    grantedCredits: 10000
   },
   proSubscriptionConfig: {
     price: 1.99,
-    grantedCredits: 500000
+    grantedCredits: 50000
   }
 }
 
 export const apiPricing = {
-  airlake: 15,
-  cruxql: 200000,
-  dwallet: 75,
-  easenft: 200,
-  frostlake: 30,
-  snowlake: 750,
-  swapstream: 75,
-  vuelock: 5,
+  airlake: 2,
+  dwallet: 5,
+  easenft: 20,
+  frostlake: 2,
+  snowlake: 100,
+  swapstream: 10,
+  vuelock: 2,
 }